+++ src/App.vue
@@ -8,8 +8,8 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import AppBar from './pages/AppBar.vue'
-import AppFooter from './pages/AppFooter.vue'
+import AppBar from './components/AppBar.vue'
+import AppFooter from './components/AppFooter.vue'
 
 export default Vue.extend({
   name: 'App',
